+++ packages/gamut/src/GridForm/GridFormInputGroup/__tests__/GridFormInputGroup-test.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 
 import {
   stubFileField,
+  stubRadioGroupField,
   stubSelectField,
   stubTextField,
   stubTextareaField,
@@ -51,6 +52,14 @@ describe('GridFormInputGroup', () => {
     });
 
     expect(wrapped.text()).toEqual(text);
+  });
+
+  it('renders a radio group when the field type is radio-group', () => {
+    const { wrapped } = renderComponent({
+      field: stubRadioGroupField,
+    });
+
+    expect(wrapped.find('input[type="radio"]')).toHaveLength(2);
   });
 
   it('renders a select when the field type is select', () => {