+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 7.6.0 (2020-05-08)
+
+* feat(Gamut): Allowed Alert to receive a class name (#784) ([fd29141](https://github.com/RyzacInc/client-modules/commit/fd29141)), closes [#784](https://github.com/RyzacInc/client-modules/issues/784)
+
+
+
+
+
+## <small>7.5.3 (2020-05-07)</small>
+
+* Allowed GridForm to not take in any fields (#787) ([a9bf324](https://github.com/RyzacInc/client-modules/commit/a9bf324)), closes [#787](https://github.com/RyzacInc/client-modules/issues/787)
+
+
+
+
+
+## <small>7.5.2 (2020-05-07)</small>
+
+* Allow GridFormCheckboxInput to take in a ReactNode as its label (#785) ([63ecaa1](https://github.com/RyzacInc/client-modules/commit/63ecaa1)), closes [#785](https://github.com/RyzacInc/client-modules/issues/785)
+
+
+
+
+
+## <small>7.5.1 (2020-05-07)</small>
+
+* Fixed prettier issue on master (#786) ([f012068](https://github.com/RyzacInc/client-modules/commit/f012068)), closes [#786](https://github.com/RyzacInc/client-modules/issues/786)
+
+
+
+
+
+## 7.5.0 (2020-05-06)
+
+* feat(Gamut): added RadioGroup inputs for GridForm (#759) ([6d969b8](https://github.com/RyzacInc/client-modules/commit/6d969b8)), closes [#759](https://github.com/RyzacInc/client-modules/issues/759)
+
+
+
+
+
+## <small>7.4.8 (2020-05-06)</small>
+
+* Removed unused deprecated button themes (#774) ([013e403](https://github.com/RyzacInc/client-modules/commit/013e403)), closes [#774](https://github.com/RyzacInc/client-modules/issues/774)
+
+
+
+
+
 ## <small>7.4.7 (2020-05-05)</small>
 
 **Note:** Version bump only for package @codecademy/gamut